backend/requirements.txt
fastapi>=0.115
python-multipart>=0.0.12
uvicorn[standard]>=0.32
anthropic>=0.40
openai>=1.50
psycopg[binary]>=3.2
psycopg-pool>=3.2
pgvector>=0.3
pydantic>=2.9
python-dotenv>=1.0
